+++ objectified-ui/public/WHATS_NEW.md
@@ -90,7 +90,8 @@ We continue to improve the platform based on your feedback with improvements and
 - Import Improvements:
   - Identify unsupported features for compatibility
   - Detects and lists all x- custom extensions
-  - Added a deprecated feature warning: Flag any deprecated constructs.
+  - Added a deprecated feature warning: Flag any deprecated constructs
+  - Added a visual preview of all schemas before import
 - Control Panel:
   - Versions page now has the ability to view the relationship graph for each version
 - OpenAPI Support: